[refactor] merge_predictions submission.py로 이전 및 run_predict import 교체
run_predict.py 인라인 함수를 submission.py로 이전해 책임 분리
